requirements.txt
torch
transformers
datasets
accelerate
peft
trl
wandb
anthropic
tqdm
pyyaml
huggingface_hub
